Starting BBox with -d or /d switch enables debug log window, piping all debug print there.
Useful for those who want to debug problems without needing to compile and debug the source code themselves. It should also (with more work) help US in the future debug problems users are having, by asking them to show us the log output.

In the future, i'll probably make this an actual windows form piggybacking off of a rich text box. This way, it's possible to turn it toggle it shown or not in the Tools menu. (and also to make it easier to deal with for end users, specifically copy paste functionality)